pyproject.toml
[tool.poetry]
name = "while-you-are-at-it-server"
version = "0.1.0"
description = ""
authors = ["thewoowon <thewoowon@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
scikit-learn = "^1.5.2"
uvicorn = "^0.32.1"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
python-decouple = "^3.8"
alembic = "^1.14.0"
aiosqlite = "^0.20.0"
greenlet = "^3.1.1"
requests = "^2.32.3"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
